// src/openapi/model/api-mypage-notification.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 * @interface ApiMypageNotification
 */
export interface ApiMypageNotification {
    /**
     *
     * @type {string}
     * @memberof ApiMypageNotification
     */
    createdAt?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageNotification
     */
    title?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageNotification
     */
    body?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageNotification
     */
    pathName?: string;
}