// src/openapi/model/api-mypage-notifications.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import {ApiMypageNotification} from './api-mypage-notification';

/**
 *
 * @export
 * @interface ApiMypageNotifications
 */
export interface ApiMypageNotifications {
    /**
     *
     * @type {number}
     * @memberof ApiMypageNotifications
     */
    total?: number;
    /**
     *
     * @type {string}
     * @memberof ApiMypageNotifications
     */
    userNotificationCheckDate?: string;
    /**
     *
     * @type {Array<ApiMypageNotification>}
     * @memberof ApiMypageNotifications
     */
    data?: Array<ApiMypageNotification>;
}